refactor(reposição): badge de baixo giro no cockpit no lugar do falso-alerta "sem parâmetro"
O aviso amarelo "N SKUs sem histórico para cálculo automático" parecia pendência de automação, mas era informativo: são itens de baixo giro sem venda recente (o auto-apply/cold-start já cobrem quem TEM dado). Gerava a dúvida "não era pra ser tudo automático?".

- Remove SmartAlertsSection (só usado no cockpit) e o substitui por um badge discreto que leva ao painel dedicado /admin/reposicao/baixo-giro.
- O badge conta o MESMO universo do painel (402 hoje), não os 172 "sem parâmetro" — senão o número mentiria vs. a lista mostrada ao clicar.
- Filtro de baixo giro extraído p/ constante única BAIXO_GIRO_OR_FILTER (fonte única entre painel e badge; evita drift silencioso).

Frontend puro (sem migration/edge). typecheck + lint + test (822) verdes.
